+++ Src/Library/changelog.md
@@ -1,2 +1,5 @@
+### NEW
+- `.UseSwaggerGen(uiConfig: c => c.DeActivateTryItOut())` extension method to de-activate the try it out button by default.
+
 ### IMPROVEMENTS
 - prevent `UseFastEndpoint()` call overwriting the DI registered `JsonOptions` [#info](https://discord.com/channels/933662816458645504/1103132906681012295)